feat: allow for a local config file repo (#2)
* Some failing tests

* Allow repo to be local path or URL

* Add failing test for specific config file commit

This was unanticipated. It turns out that git clone cannot clone a
specific commit; only tags and branches are allowed. This will need some
reworking then.

* fix: checkout version after cloning

* fix: update default repo for CLI

* Require a local name when downloading config files

* Make sure the clone is correctly checked out

* Fix bug when creating config from path

If the config was checked out at a commit and not a tag or a branch, the
command would fail. Wait is already called when calling cmd.Output, so
no need to wait at the end.