Revamp how configs are downloaded
Now there is no dependency on the old config file struct for downloading
a config. Instead I create a very basic (and actually not valid) plumber
file in the command, and a pointer to this is passed to download. This
then defines the pipeline name and version, and the remaining fields are
filled out by the download function. Not sure this is the best way to go,
but I'll stick with it for now. Will likely rework this at a later time.