Remove pipeline check
This check was to make sure that the requested pipeline exists. For a
couple of reasons this is not really necessary. Firstly, if a pipeline
is not defined in the config files, the execution will fail before the
pipeline is downloaded, and secondly, nextflow will report if the
pipeline can't be found if for some reason an invalid pipeline has been
added to the config files.

This removes some of the complexity in that we don't need to have logic
for a github token, and the problems surrounding the github rate limit.