feat: update plumberfile structure
This results in better support for both snakemake and nextflow
pipelines. The schema is also better and moves a big chunk of the
validation to the JSON schema. These changes also makes it possible to
easily increment the schema version in the future.

Some of the things that have changed:

- Refactor plumberfile validation
- Add hints in errors if files are missing
- Don't create directories if they already exist
- Don't force plumberfile version
- Schema validation now takes a slice of bytes instead of a reader
- Add path to newly parsed plumberfile
- Have Profiles() return []Profile, not []string
- Don't try to resolve paths with environment variables